requirements.txt
pandas==2.2.2
langchain==0.2.10
langchain-groq==0.1.3
python-dotenv==1.0.1
openai==1.40.1